Adopt repository
